Do not emit joined generators into optimized function bodies
Summary:
Release note: none

Joining effects results in joined generators that contain all of the I/O operations of the argument effects, but does not change the arguments.

Normally, this does not matter because only the joined generator is used. In the case of an optimized function, however, the generator tree is traversed via the effects tree and this resulted in duplicate I/O operations.

To fix this, the tree walk now ignores joined generators. Also included in this PR, some fixes to make sure that only the (joined) effects resulting from _createAdditionalEffects are used, and not the effects that went into it.
Closes https://github.com/facebook/prepack/pull/1758

Differential Revision: D7687328

Pulled By: hermanventer

fbshipit-source-id: c3a333369bc403bed5f986e5d9d59cc56ab62ced