Turn TODO into invariants
Summary:
Release Notes: None

Adds invariant to visitGenerator enforcing that if createdObjects has an entry for one of the createdObjects, it is due to a child generator. This should be true because optimized functions' nested effects' createdObjects are a strict subset of the overall optimized function effects.
Closes https://github.com/facebook/prepack/pull/1753

Differential Revision: D7710579

Pulled By: cblappert

fbshipit-source-id: e44bd899ff58498d2307a0aa6a5016cdd0da60ed