Add React-related internal regression test to Prepack
Summary:
This adds a React-specific internal test for Prepack to fbsource. It's meant to be land-blocking.

I checked in a manually edited bundle that has no dependencies except `react` and `react-dom` (so it's not too difficult to maintain). I also checked in a snapshot test for what it should render.

The test step consists of running the Jest test (to verify it works *before* Prepack), then using the existing `debug-fb-www` command to produce a bundle, and finally running the same Jest test again.

Currently this only checks the initialization path. When render path serializer bugs are fixed, we can run the same bundle in the render mode as well (and catch further regressions).

Reviewed By: NTillmann

Differential Revision: D7675812

fbshipit-source-id: 99fe3bd700ae3557ee6245f1e8e1e835399625b5