Capture prefix effects when composing pnc
Summary:
Release note: none

When composing a newly constructed PossiblyNormalCompletion with saved one, the resulting completion needs to capture the effects that happened on the normal branch, in between the saved fork point and the fork point of the newly constructed completion, because these effects also apply to the abrupt part of the new completion.

When joining up such a composed completion, the effects need to be applied (and later reversed), otherwise the nested joins might capture values as they were before the first fork, rather than as there were at the time of the nested fork.
Closes https://github.com/facebook/prepack/pull/1769

Differential Revision: D7746284

Pulled By: hermanventer

fbshipit-source-id: d2396f29f93c676e330ac6460f5ac9e638329528