Properly strips on* events in first render and fixes tests
Summary:
Release notes: fixes a regression where on* events would not be stripped in React firstRender mode

This PR fixes a recent regression in the stripping of onEvent names in React compiler's firstRender mode. The associated tests have been updated so the snapshots should pick up this regression in the future.
Closes https://github.com/facebook/prepack/pull/1774

Differential Revision: D7747737

Pulled By: trueadm

fbshipit-source-id: e0dd65a7e4d73844c499d1d8b2f9ae8046bfec2b