Forcing clearing traverse cache after each traversal.
Release notes: Use 66% less memory, run 42% faster

For a large internal benchmark, this...
- reduces peak memory usage from 9605Mb to 3307Mb,
- reduces run time from 216s to 125s.

Internal, it seems to me that the traverse cache aims at using a WeakMap.
It is not clear to me what exactly is going wrong (is some non-ES6 WeakMap
implementation kicking in?), but this effectively fixes it in any case.

I suspect that the gains in run time come from simply no longer triggering
excessive numbers of GCs due to extremely high memory usage.