Handle abstract conditional break inside definite for loop
Summary:
Release note: Allow break/continue to be guarded by abstract conditions inside loops

This addresses the first example in issue: #1764

Loops represent join points for break and continue completions. They need to be handled much like the return completions that join at function end. The main difficulty is that break and continue can specify a target. For now, we punt on that and give a compile time error.
Closes https://github.com/facebook/prepack/pull/1778

Differential Revision: D7762824

Pulled By: hermanventer

fbshipit-source-id: d8d3699277b7bb20817a27ecc24f3c503904fc74