Don't emit object property modifications to dead objects.
Release notes: None

This fixes #1771.

When an optimized function modifies properties of objects
created outside of its scope, we can omit such property modifications
if the modified object is not used.
This nicely fits in with the fixed point model of the visitor.
(The objects involved in object property modification used to complete bypass the whole visiting process, resulting in the crash.)
Note that a side effect certain double-mutation errors won't be issued to the user.

Added simple regression test.
Manually verified that this also makes the larger React test mentioned in the issue pass.