Promote empty to undefined for optimized function return values
Summary:
Release note: none

EmptyValue is a signal that the property that (might) have the value could in fact not actually be there. This matters when constructing objects at runtime, since code has to be emitted to deal with the case where the property will not be created at all, due to a runtime switch.

On the other hand, when the value of a missing property is *used*, it should be the undefined value. For that scenario, there is a promoteEmptyToUndefined on Value and this should be called before serializing empty values that are actually used.

This was not done for values returned from optimized functions. This PR plugs that hole.
Closes https://github.com/facebook/prepack/pull/1787

Differential Revision: D7776331

Pulled By: hermanventer

fbshipit-source-id: ebf8f5335e8ed3f6c557e00f3c88b08ccbd9c14d