Don't emit object property modifications to dead objects.
Summary:
Release notes: None

This fixes #1772.

When an optimized function modifies properties of objects
created outside of its scope, we can omit such property modifications
if the modified object is not used.
This nicely fits in with the fixed point model of the visitor.
(The objects involved in object property modification used to complete bypass the whole visiting process, resulting in the crash.)
Note that as a positive side effect certain double-mutation errors won't be issued to the user.

Added simple regression test.

Manually verified that this also makes the larger React test mentioned in the issue pass.

Also fixed ordering ordering of effects applications when processing queue in residual heap visitor. This bug existed for a while, but was only brought out by the existing serializer tests after the more aggressive queueing of modified object property binding visitng.
Closes https://github.com/facebook/prepack/pull/1793

Differential Revision: D7784644

Pulled By: NTillmann

fbshipit-source-id: 3dd21cc89275a30c5312f7bdd94e7d0364d02959