Simplify things by removing an unnecessary evaluateForEffects call.
Summary:
Release note: none

Resolves issue #1799

react/utils/getValueFromFunctionCall is called from within an evaluateForEffects and is expected to produce a current state that has the effects of the call applied to it. It therefore makes no sense to use evaluateForEffects within the implementation of getValueFromFunctionCall.

Making this simpler also resolves #1799. The reason for this is commented on in the code.
Closes https://github.com/facebook/prepack/pull/1805

Differential Revision: D7803343

Pulled By: hermanventer

fbshipit-source-id: e2687c60f0ce444ff4191b830c4571dcb71ed1a6