Refine type of PropertyBinding.key
Release notes: None

The type used to be just `any`; not sure what the historical story here is,
but the visitor was unaware that this in fact could be an AbstractValue as well as a string.
This meant that some keys might not get visited when they should have been.

This refines the key type and accounts for the different possible types in all referencing places.