Refine type of PropertyBinding.key
Summary:
Release notes: None

This fixes #1771.

The `PropertyBinding` `key` type used to be just `any`; not sure what the historical story here is,
but the visitor was unaware that this in fact could be an `AbstractValue` as well as a `string`.
This meant that some keys might not get visited when they should have been.

This refines the key type and accounts for the different possible types in all referencing places.
Closes https://github.com/facebook/prepack/pull/1816

Differential Revision: D7827049

Pulled By: NTillmann

fbshipit-source-id: 88fa2e036d1d4393d025af38954e309f6632a839