Generalize temporal unknown property assignments
Summary:
Release note: Support multiple assignments to unknown properties in optimized functions

Issue #1818.

The code in #1822 can only deal correctly with a single assignment. This generalizes that code to generate separate assignments for each different (key, value) pair found in the conditional value tree that is used to model a group of assignments to unknown properties of known objects.
Closes https://github.com/facebook/prepack/pull/1828

Differential Revision: D7840880

Pulled By: hermanventer

fbshipit-source-id: 8f2bb12cfd5246dc13d5a360ff9bb16a9326dd41