Follow up to UpdateExpression
It's not safe to mutate an expression in a build function. Those needs to
be emitted as statements.

We don't have to deal with the assignment part of it though. We can just
calculate the new value and deal with the fact that ToNumber might throw
or mutate.

The new value we can just pass to PutValue which takes care of the rest.