Follow up to UpdateExpression
Summary:
Follow up to #1839

It's not safe to mutate an expression in a build function. Those needs to be emitted as statements.

We don't have to deal with the assignment part of it though. We can just calculate the new value and deal with the fact that ToNumber might throw or mutate.

The new value we can just pass to PutValue which takes care of the rest.
Closes https://github.com/facebook/prepack/pull/1847

Differential Revision: D7853622

Pulled By: sebmarkbage

fbshipit-source-id: 1e370c2961048f5ce22becb7385a84b92ecc7045