Fix _WrapAbstractInObject to deal with mixed primitive types
Summary:
Release note: Fixed bug with implicitly converting unknown values to Object

Resolves #1845

Joined values can be of type PrimitiveValue, so the invariant in the default case of the switch in _WrapAbstractInObject is actually not true.

Added handling for this case, throwing an error if the joined value might be null, undefined or missing and converting it lazily to Object otherwise.
Closes https://github.com/facebook/prepack/pull/1849

Differential Revision: D7853117

Pulled By: hermanventer

fbshipit-source-id: eb8e8d85c13f9eee4320badb2209c5581f27ef0c