Deal with possibly normal completions from joined switch
Summary:
Release note: Support switch statements where the switch value is abstract and one or more cases throw

Resolves #1838

When only some of the cases in a switch statement throws, the join results in a possibly normal completion and this has to get composed with the saved completion.

The test case also revealed that the join was happening in the wrong order, so that is fixed as well.
Closes https://github.com/facebook/prepack/pull/1844

Differential Revision: D7861034

Pulled By: hermanventer

fbshipit-source-id: f4a8614e1f8769f942018345e2e1024d8013cf67