Deal with joins of joins when promoting
Summary:
Release note: none

While debugging various issues with generators, I noticed that joinEffectsAndPromoteNested does not quite live up to its advertising because it does not handle all the permutations of results that could contain abrupt completions of the given type.

Fixed this and added test cases. The tree manipulations are quite subtle. I went to some trouble to verify that they are correct, using diagrams, but did not want to put the diagrams in the code, so the reading may be a bit tough.
Closes https://github.com/facebook/prepack/pull/1873

Differential Revision: D7886038

Pulled By: hermanventer

fbshipit-source-id: df874ca5fad771932490428d54ff22b022fec74c