Do not apply the effects of JoinedAbruptCompletions.
Summary:
Release note: Fix problems with duplicated and missing statements

Resolves issue: #1842

Code duplication happened because the joined effects of two abrupt completions were applied first in aggregate and then again individually.

While debugging through this, I found it much simpler to understand Generator.fromEffects when applied to an Effects that has not been joined, so now things are closer to what they were before I started trying to hack past issues. The main difference is that a fully joined up set of effects is applied before serializing the result from Generator.fromEffects.

All of this caused an invariant to fail during a test. The invariant is complicated, fails in other new scenarios and is being reworked by Nikolai, so I've removed it for now in order to get these fixes in so that we are all on the same page again.
Closes https://github.com/facebook/prepack/pull/1903

Differential Revision: D7920610

Pulled By: hermanventer

fbshipit-source-id: 89b6ac9f1b8df776657c3a93550fdd3eb2ae4cd7