Fixes an issue with getOwnPropertyDescriptor
Summary:
Release notes: fixes an issue with getOwnPropertyDescriptor when dealing with a native prototype function

This PR fixes #1285, specifically when using `Object.getOwnPropertyDescriptor` on a native prototype function, we have to resort to returning a temporal abstract as this function isn't available at build time. I also renamed all the tests with typos in them.

cc anilanar
Closes https://github.com/facebook/prepack/pull/1913

Differential Revision: D7935215

Pulled By: trueadm

fbshipit-source-id: d7e33cf8bd73aa2c1156966ad86e9da9fb0404e4