General clean up of React reconciler + serializer code
Summary:
Release notes: none

This is a tidy up of the React reconciler and optimization code. I've removed the React logic from `serializer/functions.js` and put it all in its own module so that it's easier to debug and read the React specific code from generic Prepack optimization code. Furthermore, I took some time to rename `render` in reconciler to `resolve` as it's not actually doing any "rendering".
Closes https://github.com/facebook/prepack/pull/1968

Differential Revision: D8010381

Pulled By: trueadm

fbshipit-source-id: c61bdfec91f250f898d54623d92c20a545b965d7