Fixing binding value after havocing
Release notes: Fixing havocing of bindings.

This fixes #1973 and #2003.

When having a binding,
the binding's value at that point in time would remain in the binding,
and participate in the whole value merging process,
which was meant to compute final values and caused issues.

To fix it, an explicit assignment is now emitted at binding havocing time,
and the binding value, which is irrelevant from then on, is set to undefined.