Fixing binding value after havocing
Summary:
Release notes: Fixing havocing of bindings.

This fixes #1973 and it fixes #2003.

When havocing a binding,
the binding's value at that point in time would remain in the binding,
and participate in the whole value merging process,
which was meant to compute final values and caused issues.

To fix this, an explicit assignment is now emitted at binding havocing time with the current binding value, and afterwards, the current binding value, which is irrelevant from then on, is set to undefined.

This effectively fixes the issue for bindings, but the whole way binding havocing works is a bit inefficient (after havocing, Prepack effectively stops tracking that binding); even worse, similar problems that we observed here for bindings should exist for object properties as well; and the generated code looks a bit awful. I'll follow up creating separate issues for those, but they are out of scope for this change.

Adding regression tests.
Closes https://github.com/facebook/prepack/pull/2010

Differential Revision: D8153779

Pulled By: NTillmann

fbshipit-source-id: 9d64980e0bf2c8afe018ba0814907b90d19a9f77