Ensure react lazy initializer is in the top generator body
Summary:
Release notes: none

We need to ensure we emit the ReactElement lazy initializer is in the top generator body (the function body) of the React component render. If we emit it in a nested generator body, it may not get called in the right place after we've nested many components. This also renames a bunch of functions to optimized functions rather than additional functions.

I've got an issue tracking the progress on a follow up PR to add regression tests for this PR: https://github.com/facebook/prepack/issues/2027
Closes https://github.com/facebook/prepack/pull/2026

Differential Revision: D8172780

Pulled By: trueadm

fbshipit-source-id: 5d3457df467201b4a8ee7a0e9186fbc8ade7293b