Better support `indexOf` and `reverse` on unknown arrays with numeric properties
Summary:
Release notes: support `indexOf` and `reverse` on unknown arrays with numeric properties

This PR aims to better support `indexOf` and `reverse` on unknown arrays with numeric properties. Previously, these two methods on the array would cause a FatalError in pure mode that would recover and havoc the array itself and the arguments. I don't believe this is necessary for these two methods in pure mode because they don't mutate the arguments passed. In pure mode we expect that calling `toString` on the arguments won't affect global scope, so these operations should be fine to be emitted as temporals with them being havoced.

I also added a TODO to an area I want to look at next week that I feel will unlock many more optimizations and inlining possibilities (with regards to handling conditional abstract objects as the context of the function call).
Closes https://github.com/facebook/prepack/pull/2033

Differential Revision: D8172837

Pulled By: trueadm

fbshipit-source-id: 846e28c0e0a6d0d475ad6ba6ef705bf0eae55b85