When havocing immutable bindings, keep around and later use the leaked value
Summary:
Release notes: None

Adding regression test.
While this really just does something for const bindings that we probably don't care much about,
this is leading up to a similar change for final objects which is much more relevant.

(And eventually, the last known value of a havoced mutable binding and havoced non-final objects in between external calls should be handled in this way as well.)
Closes https://github.com/facebook/prepack/pull/2031

Differential Revision: D8172560

Pulled By: NTillmann

fbshipit-source-id: 4e839113d7e5cd0f676e0017dd774da82b6dcdcf