Add in invariant checks to ensure we fail early on impossible cases
Summary:
Release notes: none

After speaking to NTillmann, there seems to be cases where impossible cases are occuring during the React reconcilation. This PR adds in invariants so we fail early (rather than emit bad output). `type` and `props` are both parts of the immutable ReactElement. The ReactElement is final as is the props object. There should never be a case where the properties of a ReactElement are conditional abstract values – this violates so many parts of the reconciliation process.

There's only one codepath that we support for creating ReactElements, and that's via the function `createReactElement` in `react/elements.js`. The ReactElement is made final after this, so the properties of the ReactElement should never change, even to conditionals.
Closes https://github.com/facebook/prepack/pull/2037

Differential Revision: D8177989

Pulled By: trueadm

fbshipit-source-id: 4e02333c25b69c8c6a9b08b34c405db91604cf8f