Make joinPossiblyNormalCompletionWithAbruptCompletion live up to its advertised behavior
Summary:
Release note: none

joinPossiblyNormalCompletionWithAbruptCompletion said that it always returned effects with completions of type JoinedAbruptCompletions. Some callers took it at is word. Turns out it was lying.

Fixed things so that the advertising is correct. Also fixed an issue with a generator being lost during joining. (This was a latent bug uncovered by this fix.)

Refactored the code that fails when a loop body contains both break and continue completions governed by abstract conditions.
Closes https://github.com/facebook/prepack/pull/2030

Reviewed By: trueadm

Differential Revision: D8188966

Pulled By: gaearon

fbshipit-source-id: 7b74d63ee34b6492614a4386e6ad7955c96cbf78