Fix #2051 and handle all Array.prototype methods for unknown arrays
Summary:
Release notes:

This PR is a follow up to https://github.com/facebook/prepack/pull/2051, which we found to have issues and had to be reverted. The main cause of the issue in the previous PR was that `realm.evaluateWithAbstractConditional` was being given a `ReturnCompletion`, which are a type of `AbruptCompletion`. There is logic to correctly throw such completions. So instead of passing a `ReturnCompletion` to `realm.evaluateWithAbstractConditional`, we instead pass the value and then wrap the `realm.evaluateWithAbstractConditional` call with a function that adds back on the `ReturnCompletion`.

Fixing this brought about a bunch of deeper issues that were causing tests to fail (regression tests added) around the unknown array and its prototype methods. We were only handling a handful of methods before, but that was fragile. So this PR now adds support for all the of current prototype methods on `Array` for unknown arrays – following the spec as to what the return value should be in each case.

Finally, a change was made to the React reconciler to ensure that we return the unknown array in cases where it is passed an unknown array without a ReactHint (which is expected, but was missed and thus a regression from a previous PR).
Closes https://github.com/facebook/prepack/pull/2061

Differential Revision: D8229848

Pulled By: trueadm

fbshipit-source-id: 02dbd1c7e1cd03de4f6745daed57d13ea24537a8