Do not join prematurely.
Summary:
Release note: none

Quite a few of the outstanding issues arise because generators are either deleted or duplicated. After spending much time debugging such issues, I've realized that the code is not making a clear distinction between forking and joining and as a result we sometimes join when we should fork and this leads to problems.

I'm now systematically going through all the methods in Join and making sure they do the right thing in a less mysterious way. The subject of this pull request is method joinPossiblyNormalCompletionWithAbruptCompletion. The result of this method should be a just a fork because a join may not be required or desirable. The name is thus not helpful and so I've renamed it to replacePossiblyNormalCompletionWithForkedAbruptCompletion.

This broke some use cases that require the result to be joined. They had to be rewritten to use extractAndJoinCompletionsOfType to join just those parts of the fork that actually join at those code constructs.

While debugging the tests that inevitably broke while making these changes, a few other issues were uncovered that are also fixed in this PR. One of them involves specializing the result value of the abrupt completion that is being spliced into possibly normal completion (because a join is being composed into a fork).
Closes https://github.com/facebook/prepack/pull/2074

Differential Revision: D8244536

Pulled By: hermanventer

fbshipit-source-id: eded7e7233101ad5c288cd8730bf6fd3cbf82904